+++ test/test_scf/test_scf_elements.py
@@ -355,7 +355,7 @@ def test_element(dtype: torch.dtype, number: int) -> None:
     charges = torch.tensor(0.0).type(dtype)
 
     # opts["spin"] = uhf[number - 1]
-    calc = Calculator(numbers, positions, par, opts=opts)
+    calc = Calculator(numbers, positions, par, opts=opts, dtype=dtype)
     results = calc.singlepoint(numbers, positions, charges)
 
     assert pytest.approx(r, abs=tol) == results.scf.sum(-1).item()
@@ -373,7 +373,7 @@ def test_element_cation(dtype: torch.dtype, number: int) -> None:
     charges = torch.tensor(1.0).type(dtype)
 
     opts["spin"] = uhf_cation[number - 1]
-    calc = Calculator(numbers, positions, par, opts=opts)
+    calc = Calculator(numbers, positions, par, opts=opts, dtype=dtype)
 
     # no (valence) electrons
     if number in [1, 3, 11, 19, 37, 55]:
@@ -404,7 +404,7 @@ def test_element_anion(dtype: torch.dtype, number: int) -> None:
     charges = torch.tensor(-1.0).type(dtype)
 
     opts["spin"] = uhf_anion[number - 1]
-    calc = Calculator(numbers, positions, par, opts=opts)
+    calc = Calculator(numbers, positions, par, opts=opts, dtype=dtype)
     results = calc.singlepoint(numbers, positions, charges)
 
     assert pytest.approx(r, abs=tol) == results.scf.sum(-1).item()
@@ -425,7 +425,7 @@ def test_element_batch(dtype: torch.dtype, number: int, mol: str) -> None:
     charges = torch.tensor([0.0, 0.0]).type(dtype)
 
     opts["spin"] = [0, uhf[number - 1]]
-    calc = Calculator(numbers, positions, par, opts=opts)
+    calc = Calculator(numbers, positions, par, opts=opts, dtype=dtype)
     results = calc.singlepoint(numbers, positions, charges)
 
     assert torch.allclose(refs, results.scf.sum(-1), atol=tol)
